requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altgraph==0.17.4
attrs==24.2.0
discord.py==2.4.0
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
packaging==24.1
pefile==2024.8.26
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pywin32-ctypes==0.2.3
setuptools==75.1.0
yarl==1.11.1
